Initial comment fix
