--- src/main/java/org/example/movieSpring/model/Movie.java
@@ -1,27 +1,0 @@
-package org.example.movieSpring.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "movies")
-public class Movie {
-
-    @Id
-    @Column(name = "id_movie")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "movie_title")
-    private String title;
-
-    @Column(name = "movie_release_year")
-    private int release_year;
-}
