+++ src/main/java/org/example/moviespring/model/Movie.java
@@ -1,0 +1,48 @@
+package org.example.moviespring.model;
+
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "movies")
+public class Movie {
+
+    @Id
+    @Column(name = "ID")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "TITLE")
+    private String title;
+
+    @Column(name = "RELEASE_YEAR")
+    private int release_year;
+
+    public Movie() {
+    }
+
+    public Movie(Long id, String title, int release_year) {
+        this.id = id;
+        this.title = title;
+        this.release_year = release_year;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public int getRelease_year() {
+        return release_year;
+    }
+
+    public void setRelease_year(int release_year) {
+        this.release_year = release_year;
+    }
+}
